+++ test/config.example.py
@@ -2,3 +2,5 @@ host_name = "cartodb.com"
 cartodb_user = "youruser"
 api_key = "top secret"
 is_multiuser = True
+upload_file = 'path to upload file test'
+test_map = 'id for test map'