+++ widgets/ListItemWidgets.py
@@ -20,7 +20,7 @@ email                : michaelsalgado@gkudos.com, info@gkudos.com
 """
 
 from PyQt4.QtCore import Qt, QSize
-from PyQt4.QtGui import QWidget, QLabel
+from PyQt4.QtGui import QWidget, QLabel, QProgressBar
 
 from QgisCartoDB.ui.ListItem import Ui_ListItem
 
@@ -82,16 +82,26 @@ class CartoDBLayerListItem(CartoDBDatasetsListItem):
     def __init__(self, tableName=None, layer=None, size=None, rows=None):
         CartoDBDatasetsListItem.__init__(self, tableName, None, size, rows)
 
+        '''
         self.ui.statusLB = QLabel(self)
         self.ui.statusLB.setMaximumSize(QSize(100, 16777215))
         self.ui.statusLB.setAlignment(Qt.AlignCenter | Qt.AlignTrailing | Qt.AlignVCenter)
         self.ui.statusLB.setWordWrap(True)
         self.ui.horizontalLayout.insertWidget(1, self.ui.statusLB)
+        '''
+
+        self.ui.statusBar = QProgressBar(self)
+        self.ui.statusBar.setProperty("value", 0)
+        self.ui.statusBar.setFormat("Init")
+        self.ui.statusBar.hide()
+        self.ui.horizontalLayout.insertWidget(1, self.ui.statusBar)
 
         self.layer = layer
 
-    def setStatus(self, status):
-        self.ui.statusLB.setText(status)
+    def setStatus(self, status, value = 0):
+        self.ui.statusBar.setProperty("value", value)
+        self.ui.statusBar.setFormat(status)
+        self.ui.statusBar.show()
 
     def clone(self):
         return CartoDBDatasetsListItem(self.tableName, self.layer, self.size, self.rows)