+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-<a name="0.2.0"></a>
+<a name="0.2.1"></a>
 ### 0.2.1 (2015-08-19)
 
 #### Enhancements
 
 * Create maps without select layers.
+* Add status to upload items.
+* Change icons.
 
 #### Fixes